requirements.txt
requests
PyCoT
python-dateutil